crudapp_migrate/poetry-migrate/pyproject.toml
[tool.poetry]
name = "poetry-migrate"
version = "0.1.0"
description = ""
authors = ["Olegsmm2092 <radoveev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask-migrate = "^4.0.5"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
